fix: resolve all clippy warnings and test failures
- Move __mx_wrap_covenv test to impl_matrix.rs (where the type is defined)
- Fix S7 fallback test assertions: tryCatch → inherits check
- Remove unused imports (Rf_translateCharUTF8, CStr, CString)
- Remove unnecessary unsafe blocks (extract_names, scalar constructors)
- arrow_impl.rs: use .cast()/.cast_mut() instead of `as` pointer casts
- Zero clippy warnings on `--all-targets --features full`
- 271/271 macro tests pass
